feat: using new store functions in Apollon Editor index
