+++ library/lib/store/metadataStore.ts
@@ -1,9 +1,18 @@
 import { create, StoreApi, UseBoundStore } from "zustand"
 import { devtools, subscribeWithSelector } from "zustand/middleware"
 import { diagramMetadata } from "./constants"
-import { MetadataStore } from "./types"
 import { parseDiagramType } from "@/utils"
 import ydoc from "@/sync/ydoc"
+import { DiagramType } from "@/types"
+
+export type MetadataStore = {
+  diagramName: string
+  diagramType: DiagramType
+  updateDiagramName: (name: string) => void
+  updateDiagramType: (type: DiagramType) => void
+  updateMetaData: (name: string, type: DiagramType) => void
+  updateMetaDataFromYjs: () => void
+}
 
 export const createMetadataStore = (): UseBoundStore<StoreApi<MetadataStore>> =>
   create<MetadataStore>()(
@@ -24,6 +33,21 @@ export const createMetadataStore = (): UseBoundStore<StoreApi<MetadataStore>> =>
             diagramMetadata.set("diagramType", type)
           }, "store")
           set({ diagramType: type }, undefined, "updateDiagramType")
+        },
+
+        updateMetaData: (name, type) => {
+          ydoc.transact(() => {
+            diagramMetadata.set("diagramName", name)
+            diagramMetadata.set("diagramType", type)
+          }, "store")
+          set(
+            {
+              diagramName: name,
+              diagramType: type,
+            },
+            undefined,
+            "updateMetaData"
+          )
         },
 
         updateMetaDataFromYjs: () =>
